Fix compiler warnings from gcc-12 (#1316)
* Fix compiler warnings from gcc-12

* use size_t/int64_t where appropriate
* explicit cast to size_t where necessary
* initialize fields
* remove unused variables


* Revert size_t to int in both declaration and iterator in readWriteFiles


* Change char to uint8_t to match parameters in interleave()

+++ src/lib/OpenEXR/ImfMultiPartInputFile.cpp
@@ -194,7 +194,7 @@ MultiPartInputFile::getPart(int partNumber)
 const Header &
  MultiPartInputFile::header(int n) const
 {
-    if(n<0 || n >= _data->_headers.size())
+    if (n < 0 || static_cast<size_t>(n) >= _data->_headers.size ())
     {
         THROW ( IEX_NAMESPACE::ArgExc , " MultiPartInputFile::header called with invalid part " << n << " on file with " << _data->_headers.size() << " parts");
     }
@@ -830,7 +830,7 @@ bool
 MultiPartInputFile::partComplete(int part) const
 {
 
-    if(part<0 || part >= _data->_headers.size())
+    if (part < 0 || static_cast<size_t>(part) >= _data->_headers.size ())
     {
         THROW ( IEX_NAMESPACE::ArgExc , "MultiPartInputFile::partComplete called with invalid part " << part << " on file with " << _data->_headers.size() << " parts");
     }
